ai_path/requirements.txt
langchain>=0.3
langchain-openai>=0.2
langchain-community>=0.3
langchain-core>=0.3
langgraph>=0.2
tavily-python
google-search-results
requests
beautifulsoup4
lxml
streamlit>=1.35
pydantic>=2.0
python-dotenv
typing_extensions>=4.0
